文件写更新完成

+++ project/filesystem.h
@@ -51,7 +51,7 @@ struct USEROPEN
 	unsigned int fixtime;//最近修改时间
 	//char keep[1];//保留位
 	char dir[100];  //文件路径
-	unsigned int off;//读写指针
+	unsigned short off;//读写指针
 	char fcbstate;  //FCB是否被修改 1 修改 0 未修改
 	char topenfile; //打开表项是否为空 0 空 1 不空
 };
